Specify no-store for fetch cache
"no-cache" does not use new data if the last modified time
is not updated. However, we sometimes update a file by over-
writing it with an existing file, and then, if the last modified
time of the existing file is old, the last modified time of the
target file is not updated. As a result, "fetch" returns not the new
file but the old cached file.

By using "no-store", fetch always does not use cache and returns
a current file.