Fix newly added domain bug on classic Outlook appointment
RequiredAttendees has a current user even when sending a
new appointment on classic Outlook. As a result, the current
user is added to the session data for checking newly added
domains even it is a new appointment. That causes a false warning.

This patch adds a check whether an appointment is a new appointment
or not.