dialog: specify background color directly
On new Outlook, html (body) element did not cover all of a dialog, so specifying "background"
to elements did not affect to the area not covered with html (body) element.
In order to resolve that, we need to specify width and height 100% after resize in order to
hide the original background color.