Specify displayInIframe = true on displayDialogAsync
When we edit a mail on a new window on Outlook on the web
with Microsoft Edge or Google Chrome, the new dialog URL
is about:blank, and we can not allow to popup from that
URL (about:blank). In order to work displayDialogAsync
in such case, we should specify displayInIframe = true.

This patch somewhat affects to a dialog appearance.